test case and check function added

+++ service/routes.py
@@ -145,10 +145,6 @@ class WishlistResource(Resource):
         This endpoint will return a Wishlist based on it's id
         """
         app.logger.info("Request to Retrieve a wishlist with id [%s]", wishlist_id)
-        if not wishlist_id.isnumeric():
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
-        if int(wishlist_id) < 0:
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find(wishlist_id)
         if not wishlist:
             abort(status.HTTP_404_NOT_FOUND, "Wishlist with id '{}' was not found.".format(wishlist_id))
@@ -169,11 +165,9 @@ class WishlistResource(Resource):
 
         This endpoint will update a Wishlist based the body that is posted
         """
-        if not wishlist_id.isnumeric():
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
-        if int(wishlist_id) < 0:
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         app.logger.info('Request to Update a wishlist with id [%s]', wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         check_content_type("application/json")
         wishlist = Wishlist.find(wishlist_id)
         if not wishlist:
@@ -201,11 +195,9 @@ class WishlistResource(Resource):
 
         This endpoint will delete a Wishlist based the id specified in the path
         """
-        if not wishlist_id.isnumeric():
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
-        if int(wishlist_id) < 0:
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         app.logger.info('Request to Delete a wishlist with id [%s]', wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find(wishlist_id)
         if wishlist:
             wishlist.delete()
@@ -260,6 +252,8 @@ class WishlistCollection(Resource):
         wishlist = Wishlist()
         app.logger.debug('Payload = %s', api.payload)
         wishlist.deserialize(api.payload)
+        if not Wishlist.correct_type_or_400(wishlist.customer_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Customer ID needs to be a positive integer.")
         wishlist.create()
         app.logger.info('Wishlist with new id [%s] created!', wishlist.id)
         location_url = api.url_for(WishlistResource, wishlist_id=wishlist.id, _external=True)
@@ -366,9 +360,7 @@ class ProductsCollections(Resource):
     def post(self,wishlist_id):
         # ADD A ITEM TO AN WISHLIST
         app.logger.info("Request to add an item to an wishlist")
-        if not wishlist_id.isnumeric():
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
-        if int(wishlist_id) < 0:
+        if not Wishlist.correct_type_or_400(wishlist_id):
             abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find_or_404(wishlist_id)
         app.logger.debug('Payload = %s', api.payload)
@@ -392,9 +384,7 @@ class ProductsCollections(Resource):
     def get(self, wishlist_id):
         """Returns all of items of a wishlist"""
         app.logger.info("Request for Wishlist Products...")
-        if not wishlist_id.isnumeric():
-            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
-        if int(wishlist_id) < 0:
+        if not Wishlist.correct_type_or_400(wishlist_id):
             abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find_or_404(wishlist_id)
         results = [product.serialize() for product in wishlist.products]
@@ -431,6 +421,8 @@ class ProductsResource(Resource):
         """
         app.logger.info(
             "Request to get an item with id: %s from wishlist with id: %s", item_id, wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         wishlist = Wishlist.find_or_404(wishlist_id)
         product = Product.find_or_404(item_id)
         return product.serialize(), status.HTTP_200_OK
@@ -443,6 +435,8 @@ class ProductsResource(Resource):
         """
         app.logger.info(
             "Request to delete product with id: %s from wishlist with id: %s", item_id, wishlist_id)
+        if not Wishlist.correct_type_or_400(wishlist_id):
+            abort(status.HTTP_400_BAD_REQUEST, "Wishlist ID needs to be a positive integer.")
         product = Product.find(item_id)
         if product:
             product.delete()
